Crate rusoto_macie

Amazon Macie
Amazon Macie is a security service that uses machine learning to automatically discover, classify, and protect sensitive data in AWS. Macie recognizes sensitive data such as personally identifiable information (PII) or intellectual property, and provides you with dashboards and alerts that give visibility into how this data is being accessed or moved. For more information, see the Macie User Guide.

If you're using the service, you're probably looking for MacieClient and Macie.

Structs

AssociateMemberAccountRequest

AssociateS3ResourcesRequest

AssociateS3ResourcesResult

ClassificationType
The classification type that Amazon Macie applies to the associated S3 resources.

ClassificationTypeUpdate
The classification type that Amazon Macie applies to the associated S3 resources. At least one of the classification types (oneTime or continuous) must be specified.

DisassociateMemberAccountRequest

DisassociateS3ResourcesRequest

DisassociateS3ResourcesResult

FailedS3Resource
Includes details about the failed S3 resources.

ListMemberAccountsRequest

ListMemberAccountsResult

ListS3ResourcesRequest

ListS3ResourcesResult

MacieClient
A client for the Amazon Macie API.

MemberAccount
Contains information about the Amazon Macie member account.

S3Resource
Contains information about the S3 resource. This data type is used as a request parameter in the DisassociateS3Resources action and can be used as a response parameter in the AssociateS3Resources and UpdateS3Resources actions.

S3ResourceClassification
The S3 resources that you want to associate with Amazon Macie for monitoring and data classification. This data type is used as a request parameter in the AssociateS3Resources action and a response parameter in the ListS3Resources action.

S3ResourceClassificationUpdate
The S3 resources whose classification types you want to update. This data type is used as a request parameter in the UpdateS3Resources action.

UpdateS3ResourcesRequest

UpdateS3ResourcesResult

Enums

AssociateMemberAccountError
Errors returned by AssociateMemberAccount

AssociateS3ResourcesError
Errors returned by AssociateS3Resources

DisassociateMemberAccountError
Errors returned by DisassociateMemberAccount

DisassociateS3ResourcesError
Errors returned by DisassociateS3Resources

ListMemberAccountsError
Errors returned by ListMemberAccounts

ListS3ResourcesError
Errors returned by ListS3Resources

UpdateS3ResourcesError
Errors returned by UpdateS3Resources

Traits

Macie
Trait representing the capabilities of the Amazon Macie API. Amazon Macie clients implement this trait.