Trait rusoto_macie::Macie

pub trait Macie {
    fn associate_member_account(
        &self, 
        input: AssociateMemberAccountRequest
    ) -> RusotoFuture<(), AssociateMemberAccountError>;

    fn associate_s3_resources(
        &self, 
        input: AssociateS3ResourcesRequest
    ) -> RusotoFuture<AssociateS3ResourcesResult, AssociateS3ResourcesError>;

    fn disassociate_member_account(
        &self, 
        input: DisassociateMemberAccountRequest
    ) -> RusotoFuture<(), DisassociateMemberAccountError>;

    fn disassociate_s3_resources(
        &self, 
        input: DisassociateS3ResourcesRequest
    ) -> RusotoFuture<DisassociateS3ResourcesResult, DisassociateS3ResourcesError>;

    fn list_member_accounts(
        &self, 
        input: ListMemberAccountsRequest
    ) -> RusotoFuture<ListMemberAccountsResult, ListMemberAccountsError>;

    fn list_s3_resources(
        &self, 
        input: ListS3ResourcesRequest
    ) -> RusotoFuture<ListS3ResourcesResult, ListS3ResourcesError>;

    fn update_s3_resources(
        &self, 
        input: UpdateS3ResourcesRequest
    ) -> RusotoFuture<UpdateS3ResourcesResult, UpdateS3ResourcesError>;
}

Trait representing the capabilities of the Amazon Macie API. Amazon Macie clients implement this trait.

Required methods

fn associate_member_account(
&self,
input: AssociateMemberAccountRequest
) -> RusotoFuture<(), AssociateMemberAccountError>

Associates a specified AWS account with Amazon Macie as a member account.

fn associate_s3_resources(
&self,
input: AssociateS3ResourcesRequest
) -> RusotoFuture<AssociateS3ResourcesResult, AssociateS3ResourcesError>

Associates specified S3 resources with Amazon Macie for monitoring and data classification. If memberAccountId isn't specified, the action associates specified S3 resources with Macie for the current master account. If memberAccountId is specified, the action associates specified S3 resources with Macie for the specified member account.

fn disassociate_member_account(
&self,
input: DisassociateMemberAccountRequest
) -> RusotoFuture<(), DisassociateMemberAccountError>

Removes the specified member account from Amazon Macie.

fn disassociate_s3_resources(
&self,
input: DisassociateS3ResourcesRequest
) -> RusotoFuture<DisassociateS3ResourcesResult, DisassociateS3ResourcesError>

Removes specified S3 resources from being monitored by Amazon Macie. If memberAccountId isn't specified, the action removes specified S3 resources from Macie for the current master account. If memberAccountId is specified, the action removes specified S3 resources from Macie for the specified member account.

fn list_member_accounts(
&self,
input: ListMemberAccountsRequest
) -> RusotoFuture<ListMemberAccountsResult, ListMemberAccountsError>

Lists all Amazon Macie member accounts for the current Amazon Macie master account.

fn list_s3_resources(
&self,
input: ListS3ResourcesRequest
) -> RusotoFuture<ListS3ResourcesResult, ListS3ResourcesError>

Lists all the S3 resources associated with Amazon Macie. If memberAccountId isn't specified, the action lists the S3 resources associated with Amazon Macie for the current master account. If memberAccountId is specified, the action lists the S3 resources associated with Amazon Macie for the specified member account.

fn update_s3_resources(
&self,
input: UpdateS3ResourcesRequest
) -> RusotoFuture<UpdateS3ResourcesResult, UpdateS3ResourcesError>

Updates the classification types for the specified S3 resources. If memberAccountId isn't specified, the action updates the classification types of the S3 resources associated with Amazon Macie for the current master account. If memberAccountId is specified, the action updates the classification types of the S3 resources associated with Amazon Macie for the specified member account.

Loading content...